+++ R/stabilize_dbl.R
@@ -163,8 +163,8 @@ stabilise_double_scalar <- stabilize_dbl_scalar
   x_arg = caller_arg(x),
   call = caller_env()
 ) {
-  min_value <- to_dbl_scalar(min_value, call = call)
-  max_value <- to_dbl_scalar(max_value, call = call)
+  min_value <- to_dbl_scalar(min_value, allow_null = TRUE, call = call)
+  max_value <- to_dbl_scalar(max_value, allow_null = TRUE, call = call)
   min_failure_locations <- .find_failures(x, min_value, `<`)
   max_failure_locations <- .find_failures(x, max_value, `>`)
   if (is.null(min_failure_locations) && is.null(max_failure_locations)) {