+++ R/to_dbl.R
@@ -145,8 +145,8 @@ to_dbl.default <- function(x, ..., x_arg = caller_arg(x), call = caller_env()) {
 to_dbl_scalar <- function(
   x,
   ...,
-  allow_null = TRUE,
-  allow_zero_length = TRUE,
+  allow_null = FALSE,
+  allow_zero_length = FALSE,
   x_arg = caller_arg(x),
   call = caller_env(),
   x_class = object_type(x)