+++ R/to_fct.R
@@ -203,8 +203,8 @@ to_fct.default <- function(
 to_fct_scalar <- function(
   x,
   ...,
-  allow_null = TRUE,
-  allow_zero_length = TRUE,
+  allow_null = FALSE,
+  allow_zero_length = FALSE,
   levels = NULL,
   to_na = character(),
   x_arg = caller_arg(x),