+++ R/to_int.R
@@ -137,8 +137,8 @@ to_int.default <- function(x, ..., x_arg = caller_arg(x), call = caller_env()) {
 to_int_scalar <- function(
   x,
   ...,
-  allow_null = TRUE,
-  allow_zero_length = TRUE,
+  allow_null = FALSE,
+  allow_zero_length = FALSE,
   x_arg = caller_arg(x),
   call = caller_env(),
   x_class = object_type(x)