+++ tests/testthat/test-to_chr.R
@@ -156,7 +156,7 @@ test_that("to_chr_scalar() allows length-1 chrs through (#22)", {
     to_chr_scalar("a"),
     "a"
   )
-  expect_null(to_chr_scalar(NULL))
+  expect_null(to_chr_scalar(NULL, allow_null = TRUE))
 })
 
 test_that("to_chr_scalar() errors for non-scalars (#22)", {
@@ -194,15 +194,15 @@ test_that("to_chr_scalar() errors for uncoerceable types (#22)", {
 test_that("to_chr_scalar() respects allow_null (#22)", {
   given <- NULL
   expect_error(
-    to_chr_scalar(given, allow_null = FALSE),
+    to_chr_scalar(given),
     class = .compile_dash("stbl", "error", "bad_null")
   )
   expect_snapshot(
-    to_chr_scalar(given, allow_null = FALSE),
+    to_chr_scalar(given),
     error = TRUE
   )
   expect_snapshot(
-    wrapped_to_chr_scalar(given, allow_null = FALSE),
+    wrapped_to_chr_scalar(given),
     error = TRUE
   )
 })
@@ -210,11 +210,11 @@ test_that("to_chr_scalar() respects allow_null (#22)", {
 test_that("to_chr_scalar respects allow_zero_length (#22)", {
   given <- character()
   expect_error(
-    to_chr_scalar(given, allow_zero_length = FALSE),
+    to_chr_scalar(given),
     class = .compile_dash("stbl", "error", "bad_empty")
   )
   expect_snapshot(
-    to_chr_scalar(given, allow_zero_length = FALSE),
+    to_chr_scalar(given),
     error = TRUE
   )
 })