+++ tests/testthat/test-to_fct.R
@@ -122,11 +122,11 @@ test_that("to_fct_scalar() provides informative error messages (#62)", {
 test_that("to_fct_scalar respects allow_zero_length (#62)", {
   given <- factor()
   expect_error(
-    to_fct_scalar(given, allow_zero_length = FALSE),
+    to_fct_scalar(given),
     class = .compile_dash("stbl", "error", "bad_empty")
   )
   expect_snapshot(
-    to_fct_scalar(given, allow_zero_length = FALSE),
+    to_fct_scalar(given),
     error = TRUE
   )
 })